backend/requirements.txt
Flask==3.1.1
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.43
requests==2.32.4
openai==1.99.9
azure-cognitiveservices-speech==1.45.0
azure-ai-textanalytics==5.3.0
python-dotenv==1.1.1

# Production dependencies
gunicorn==21.2.0
psycopg2-binary==2.9.9